pyproject.toml
[tool.poetry]
name = "diplom-10"
version = "0.1.0"
description = ""
authors = ["mustafanysan <mustafa.nysan@mail.ru>"]
readme = "README.md"
packages = [{include = "diplom_10"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
django-extensions = "^3.2.1"
pydantic = "^1.10.6"
requests = "^2.28.2"
werkzeug = "^2.2.3"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
